+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleEventsTests.java
@@ -894,19 +894,19 @@ public class ModuleEventsTests {
         Countly countly = new Countly().init(countlyConfig);
 
         countly.events().recordEvent("TEST");
-        validateEventInRQ("TEST", 0, 1, "_CLY_", "_CLY_", null);
+        validateEventInRQ("TEST", 0, 1, "_CLY_", "_CLY_", null, null);
 
         countly.views().startView("View1");
         countly.events().recordEvent("TEST1");
 
         ModuleViewsTests.validateView("View1", 0.0, 1, 3, true, true, TestUtils.map(), "_CLY_", "_CLY_", null);
-        validateEventInRQ("TEST1", 2, 3, "_CLY_", "_CLY_", null);
+        validateEventInRQ("TEST1", 2, 3, "_CLY_", "_CLY_", null, null);
 
         countly.views().startView("View2");
         countly.events().recordEvent("TEST2");
 
         ModuleViewsTests.validateView("View2", 0.0, 3, 5, false, true, TestUtils.map(), "_CLY_", "_CLY_", null);
-        validateEventInRQ("TEST2", 4, 5, "_CLY_", "_CLY_", null);
+        validateEventInRQ("TEST2", 4, 5, "_CLY_", "_CLY_", null, null);
     }
 
     /**
@@ -925,25 +925,28 @@ public class ModuleEventsTests {
         Countly countly = new Countly().init(countlyConfig);
 
         countly.events().recordEvent("TEST");
-        validateEventInRQ("TEST", 0, 1, "_CLY_", "_CLY_", "");
+        validateEventInRQ("TEST", 0, 1, "_CLY_", "_CLY_", "", "");
 
         countly.views().startView("View1");
         countly.events().recordEvent("TEST1");
 
         ModuleViewsTests.validateView("View1", 0.0, 1, 3, true, true, TestUtils.map(), "_CLY_", "_CLY_", "");
-        validateEventInRQ("TEST1", 2, 3, "_CLY_", "_CLY_", "TEST");
+        validateEventInRQ("TEST1", 2, 3, "_CLY_", "_CLY_", "TEST", "View1");
 
         countly.views().startView("View2");
         countly.events().recordEvent("TEST2");
 
         ModuleViewsTests.validateView("View2", 0.0, 3, 5, false, true, TestUtils.map(), "_CLY_", "_CLY_", "View1");
-        validateEventInRQ("TEST2", 4, 5, "_CLY_", "_CLY_", "TEST1");
+        validateEventInRQ("TEST2", 4, 5, "_CLY_", "_CLY_", "TEST1", "View2");
     }
 
-    protected static void validateEventInRQ(String eventName, int idx, int rqCount, String previousViewId, String currentViewId, String previousEventName) throws JSONException {
+    protected static void validateEventInRQ(String eventName, int idx, int rqCount, String previousViewId, String currentViewId, String previousEventName, String currentViewName) throws JSONException {
         Map<String, Object> segmentation = TestUtils.map();
         if (previousEventName != null) {
             segmentation.put("cly_pen", previousEventName);
+        }
+        if (currentViewName != null) {
+            segmentation.put("cly_cvn", currentViewName);
         }
         validateEventInRQ(TestUtils.commonDeviceId, eventName, segmentation, 1, 0.0, 0.0, "_CLY_", previousViewId, currentViewId, "_CLY_", idx, rqCount, 0, 1);
     }